requirements/dev.txt
# Core dependencies
-r base.txt

# Development tools
autopep8>=2.3.0,<3.0.0
bandit>=1.7.5,<2.0.0
black>=25.1.0,<26.0.0

# Type stubs
django-stubs>=5.2.0,<6.0.0
django-stubs-ext>=5.2.0,<6.0.0
djlint>=1.36.0,<2.0.0
filelock>=3.16.0,<4.0.0
ipython>=9.2.0,<10.0.0
isort>=6.0.0,<7.0.0
matplotlib-stubs>=0.2.0,<1.0.0
mypy>=1.15.0,<2.0.0
pandas-stubs>=2.2.3.250308
pre-commit>=4.2.0,<5.0.0
psutil>=6.1.0,<7.0.0

# Linting and formatting
pycodestyle>=2.13.0,<3.0.0
pydantic>=2.6.0,<3.0.0
pylint>=3.3.0,<4.0.0
pylint-django>=2.6.0,<3.0.0

# Testing dependencies
pytest>=8.3.0,<9.0.0
pytest-cov>=6.1.0,<7.0.0
pytest-django>=4.11.0,<5.0.0
pytest-stub>=1.1.0,<2.0.0
pytest-xdist>=3.6.0,<4.0.0
ruff>=0.11.0,<1.0.0
safety>=3.3.0,<4.0.0
types-jinja2>=2.11.9,<3.0.0
types-Pillow>=10.2.0.20240822
types-protobuf>=6.30.2.20250506
types-python-dateutil>=2.9.0.20241206
types-pytz>=2025.2.0.20250326
types-PyYAML>=6.0.12.20250402
types-requests>=2.32.0.20250328
types-setuptools>=80.3.0,<81.0.0
